Added Garden Ants
Just got the Garden Ants working in game. Most work was done on the animation.

It currently has a placeholder AI which consist of some rudimentary and some unnecessary goals. They are not able to breed but do follow the player if they're holding almost any form of sweets in their hand.